Add packaging via pyproject and editable install docs
