+++ pyproject.toml
@@ -1,0 +1,33 @@
+[build-system]
+requires = ["setuptools>=61", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "vigapp"
+version = "0.1.0"
+description = "Aplicación de diseño de vigas NTP E.060"
+readme = "README.md"
+license = {file = "LICENSE"}
+requires-python = ">=3.8"
+
+# Dependencies from requirements.txt
+dependencies = [
+    "PyQt5",
+    "matplotlib",
+    "numpy",
+    "scipy",
+    "mplcursors",
+    "pyqtgraph",
+    "sympy",
+    "python-docx",
+    "pylatex",
+    "Pillow"
+]
+
+[project.scripts]
+vigapp = "vigapp.main:main"
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["vigapp*", "src*"]
+
